+++ app/assets/javascripts/Components/groups_manager.jsx
@@ -3,8 +3,10 @@ import {render} from 'react-dom';
 
 import {withSelection, CheckboxTable} from './markus_with_selection_hoc';
 import ExtensionModal from './Modals/extension_modal';
+import {durationSort} from "./Helpers/table_helpers";
 
 class GroupsManager extends React.Component {
+
   constructor(props) {
     super(props);
     this.state = {
@@ -169,7 +171,16 @@ class GroupsManager extends React.Component {
     })
   };
 
+  extraModalInfo = () => {
+    // Render extra modal info for timed assignments only
+    if (this.props.timed) {
+      return I18n.t('assignments.timed.modal_current_duration', { duration: this.props.current_duration });
+    }
+  };
+
   render() {
+    const times = !!this.props.timed ? ['hours', 'minutes'] : ['weeks', 'days', 'hours'];
+    const title = !!this.props.timed ? I18n.t('groups.duration_extension') : I18n.t('groups.due_date_extension');
     return (
       <div>
         <GroupsActionBox
@@ -201,6 +212,8 @@ class GroupsManager extends React.Component {
               scanned_exam={this.props.scanned_exam}
               assignment_id={this.props.assignment_id}
               onExtensionModal={this.handleShowModal}
+              extensionColumnHeader={title}
+              times={times}
             />
           </div>
         </div>
@@ -210,11 +223,15 @@ class GroupsManager extends React.Component {
           weeks={this.state.selected_extension_data.weeks}
           days={this.state.selected_extension_data.days}
           hours={this.state.selected_extension_data.hours}
+          minutes={this.state.selected_extension_data.minutes}
           note={this.state.selected_extension_data.note}
           penalty={this.state.selected_extension_data.apply_penalty}
           grouping_id={this.state.selected_extension_data.grouping_id}
           extension_id={this.state.selected_extension_data.id}
           updating={this.state.updating_extension}
+          times={times}
+          title={title}
+          extra_info={this.extraModalInfo()}
           key={this.state.selected_extension_data.id} // this causes the ExtensionModal to be recreated if this value changes
         />
       </div>
@@ -334,20 +351,23 @@ class RawGroupsTable extends React.Component {
       sortable: false
     },
     {
-      Header: I18n.t('groups.extension'),
+      Header: this.props.extensionColumnHeader,
       accessor: 'extension',
       show: !this.props.scanned_exam,
       Cell: row => {
-        let extension = ['weeks', 'days', 'hours'].map(
+        let extension = this.props.times.map(
           (key) => {
-            if (row.original.extension[key]) {
+            const val = row.original.extension[key];
+            if (val) {
               // don't build these strings dynamically or they will be missed by the i18n-tasks checkers.
               if (key === 'weeks') {
-                return I18n.t('extensions.durations.weeks', {count: row.original.extension[key]});
+                return `${val} ${I18n.t('durations.weeks', {count: val})}`;
               } else if (key === 'days') {
-                return I18n.t('extensions.durations.days', {count: row.original.extension[key]});
+                return `${val} ${I18n.t('durations.days', {count: val})}`;
               } else if (key === 'hours') {
-                return I18n.t('extensions.durations.hours', {count: row.original.extension[key]});
+                return `${val} ${I18n.t('durations.hours', {count: val})}`;
+              } else if (key === 'minutes') {
+                return `${val} ${I18n.t('durations.minutes', {count: val})}`;
               } else {
                 return '';
               }
@@ -369,17 +389,7 @@ class RawGroupsTable extends React.Component {
         }
       },
       filterable: false,
-      sortMethod: (a, b) =>  {
-        a = [a.weeks || -1, a.days || -1, a.hours || -1];
-        b = [b.weeks || -1, b.days || -1, b.hours || -1];
-        if (a < b) {
-          return 1
-        } else if (b < a) {
-          return -1
-        } else {
-          return 0
-        }
-      }
+      sortMethod: durationSort
     }
   ];