+++ app/policies/grouping_policy.rb
@@ -60,4 +60,18 @@ class GroupingPolicy < ApplicationPolicy
   def no_extension?
     record.extension.nil?
   end
+
+  def view_file_manager?
+    user.student? &&
+      !(record.assignment.scanned_exam? ||
+        record.assignment.is_peer_review? ||
+        (record.assignment.is_timed? && record.start_time.nil?))
+  end
+
+  def start_timed_assignment?
+    user.student? &&
+      record.start_time.nil? &&
+      !record.past_collection_date? &&
+      record.past_assessment_start_time?
+  end
 end