+++ db/migrate/20200608190551_add_started_assignment_to_grouping.rb
@@ -1,0 +1,5 @@
+class AddStartedAssignmentToGrouping < ActiveRecord::Migration[6.0]
+  def change
+    add_column :groupings, :start_time, :datetime, null: true, default: nil
+  end
+end